[refactor] Move aot_module files
