+++ taichi/backends/vulkan/aot_module_loader_impl.h
@@ -7,7 +7,7 @@
 #include "taichi/backends/vulkan/runtime.h"
 #include "taichi/codegen/spirv/kernel_utils.h"
 
-#include "taichi/program/aot_module.h"
+#include "taichi/aot/module_loader.h"
 
 namespace taichi {
 namespace lang {
