+++ taichi/backends/wasm/aot_module_builder_impl.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "taichi/program/aot_module.h"
+#include "taichi/aot/module_builder.h"
 #include "taichi/program/kernel.h"
 #include "taichi/llvm/llvm_fwd.h"
