+++ taichi/program/program_impl.h
@@ -1,11 +1,12 @@
 #pragma once
+
+#include "taichi/aot/module_builder.h"
 #include "taichi/ir/statements.h"
 #include "taichi/system/memory_pool.h"
 #include "taichi/common/logging.h"
 #include "taichi/struct/snode_tree.h"
 #include "taichi/program/snode_expr_utils.h"
 #include "taichi/program/kernel_profiler.h"
-#include "taichi/program/aot_module.h"
 #include "taichi/backends/device.h"
 
 namespace taichi {
